few fixes (#158)
* few fixes

* removing HIPSPARSE_EXPORT in cpp file

* client: transpose_csr index types templated

* more HIPSPARSE_EXPORT in cpp file

* bump version

[ROCm/hipSPARSE commit: 6a3bca43c9835db4c321953d14bda7cd97739050]

+++ clients/include/unit.hpp
@@ -45,9 +45,9 @@
 // case
 // a wrapper will cause the loop keep going
 template <typename T>
-void unit_check_general(int M, int N, int lda, T* hCPU, T* hGPU);
+void unit_check_general(int64_t M, int64_t N, int64_t lda, T* hCPU, T* hGPU);
 
 template <typename T>
-void unit_check_near(int M, int N, int lda, T* hCPU, T* hGPU);
+void unit_check_near(int64_t M, int64_t N, int64_t lda, T* hCPU, T* hGPU);
 
 #endif // UNIT_HPP
